% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem.R
\name{rem}
\alias{rem}
\title{Random Encounter Model function}
\usage{
rem(dat, tm, v)
}
\arguments{
\item{dat}{= the data frame}

\item{tm}{= camera hours}

\item{v}{= distance travelled in 24 hours (km)}
}
\description{
This function allows the automated calculation of animal densities.km2 using camera trap data and Random Encounter Modelling.
}
\details{
The function assumes that the first 5 columns are: 
1) site number
2) camera number
3) number of individuals in each capture
4) distance to animal (in metres)
5) angle of detection (in radians).
}
\examples{

Define tm and v:

tm <- 3600
v <- 1.4

For one survey site:

rem(remDat, tm, v)

For multiple survey sites, assuming tm and v are constant:
rem(dat = grpDat[[1]], tm, v)
}
\keyword{density,}
\keyword{encounter}
\keyword{model}
\keyword{population,}
\keyword{random}

