% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/res_dat.R
\name{res_dat}
\alias{res_dat}
\title{Data Splitting function}
\usage{
res_dat(.df)
}
\arguments{
\item{df}{= the data frame}
}
\description{
Use known row number to tell function how many data to resample in each iteration
}
\examples{

res_dat <- function(.df) {
tapply(.df[,4], .df[,1], length)
}

grpSize <- res_dat(dat)
}

