% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_dat.R
\name{split_dat}
\alias{split_dat}
\title{Data Splitting function}
\usage{
split_dat(.df)
}
\arguments{
\item{.df}{= the data frame}
}
\description{
Separate data into groups prior to bootstrapping
}
\examples{

split_dat <- function(.df) {
  lapply(unique(.df[[1]]), grpDat_func, df = .df)
}

grpDat <- split_dat(dat)
}

